% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BregSurv-package.R
\docType{package}
\name{BregSurv}
\alias{BregSurv-package}
\alias{BregSurv}
\title{BregSurv: Transfer Learning for Cox Models via Bregman Divergence}
\description{
Transfer Learning for Time-to-Event Modelling via Bregman Divergence. The package supports both low-dimensional and high-dimensional settings to leverage external summary statistics or auxiliary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://um-kevinhe.github.io/BregSurv/}
}

}
\author{
\strong{Maintainer}: Yubo Shao \email{ybshao@umich.edu}

}
\keyword{internal}
