% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cox_MDTL_enet}
\alias{plot.cox_MDTL_enet}
\title{Plot Validation Results for cox_MDTL_enet Object}
\usage{
\method{plot}{cox_MDTL_enet}(
  x,
  test_z = NULL,
  test_time = NULL,
  test_delta = NULL,
  test_stratum = NULL,
  criteria = c("loss", "CIndex"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"cox_MDTL_enet"}.}

\item{test_z}{Matrix of test covariates.}

\item{test_time}{Vector of test survival times.}

\item{test_delta}{Vector of test status indicators.}

\item{test_stratum}{Vector of test strata.}

\item{criteria}{Metric to plot: \code{"loss"} or \code{"CIndex"}.}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the validation performance against \code{lambda} for MDTL elastic net estimates.
}
\examples{
\dontrun{
data(ExampleData_highdim)
train_dat_highdim <- ExampleData_highdim$train
test_dat_highdim <- ExampleData_highdim$test
beta_external_highdim <- ExampleData_highdim$beta_external

cox_MDTL_enet_est <- cox_MDTL_enet(z = train_dat_highdim$z,
                                   delta = train_dat_highdim$status,
                                   time = train_dat_highdim$time,
                                   beta = beta_external_highdim,
                                   Q = NULL,
                                   eta = 0)

plot.cox_MDTL_enet(cox_MDTL_enet_est,
                   test_z = test_dat_highdim$z,
                   test_time = test_dat_highdim$time,
                   test_delta = test_dat_highdim$status,
                   test_stratum = test_dat_highdim$stratum,
                   criteria = "CIndex")
}
}
