% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{plot.variable_importance}
\alias{plot.variable_importance}
\title{Plot Variable Importance (Selection Frequency)}
\usage{
\method{plot}{variable_importance}(
  x,
  threshold = 1,
  top = length(x$freq),
  title = "Top variables by selection frequency",
  ...
)
}
\arguments{
\item{x}{An object of class \code{variable_importance}.}

\item{threshold}{Numeric between 0 and 1. Default is 1, meaning no threshold filtering.
If not equal to 1, variables with \code{SelectionFreq > threshold} are plotted.}

\item{top}{Integer. Maximum number of variables to plot. Default is all variables.}

\item{title}{Character. Plot title. Default "Top variables by selection frequency".}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots selection frequencies as a horizontal bar chart, ordered from highest to lowest.
By default (threshold = 1), all variables are shown. If threshold != 1, only variables
with selection frequency > threshold are shown. If the number of displayed variables
exceeds \code{top}, only the top \code{top} variables are plotted.
}
