% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cox_MDTL}
\alias{plot.cox_MDTL}
\title{Plot Validation Results for Cox_MDTL Object}
\usage{
\method{plot}{cox_MDTL}(
  x,
  test_z = NULL,
  test_time = NULL,
  test_delta = NULL,
  test_stratum = NULL,
  criteria = c("loss", "CIndex"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"Cox_MDTL"}.}

\item{test_z}{Matrix of test covariates.}

\item{test_time}{Vector of test survival times.}

\item{test_delta}{Vector of test status indicators.}

\item{test_stratum}{Vector of test strata.}

\item{criteria}{Metric to plot: \code{"loss"} or \code{"CIndex"}.}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the validation performance against \code{eta} for MDTL estimates.
Compares the "Integrated" estimator (solid line) against the "Internal" baseline (dotted line, eta=0).
}
\examples{
\dontrun{
data(ExampleData_lowdim)
train_dat_lowdim <- ExampleData_lowdim$train
test_dat_lowdim <- ExampleData_lowdim$test
beta_external_lowdim <- ExampleData_lowdim$beta_external_fair

eta_list <- generate_eta(method = "exponential", n = 50, max_eta = 5)
cox_MDTL_est <- cox_MDTL(z = train_dat_lowdim$z,
                         delta = train_dat_lowdim$status,
                         time = train_dat_lowdim$time,
                         beta = beta_external_lowdim,
                         Q = NULL,
                         etas = eta_list)

plot.cox_MDTL(cox_MDTL_est,
              test_z = test_dat_lowdim$z,
              test_time = test_dat_lowdim$time,
              test_delta = test_dat_lowdim$status,
              test_stratum = test_dat_lowdim$stratum,
              criteria = "CIndex")
}
}
